requirements.txt
alembic==1.2.1
blinker==1.4
cffi==1.14.0
Click==7.0
coverage==5.1
cryptography==2.9.2
Flask==1.1.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.20
pygame==1.9.4
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
WTForms==2.2.1
